updated seat numbers to only accepts integers

+++ lib/home/firebase_post_ride.dart
@@ -9,7 +9,7 @@ Future<void> postRideToFirestore(
     String to,
     double price,
     String car,
-    double seats,
+    int seats,
     DateTime selectedDate,
     TimeOfDay selectedTime,
     context) async {
